refactor: Member controller, dto, type 패키지 이동 및 수정 (#15)
