--- src/main/java/com/octagram/pollet/member/domain/model/AuthProvider.java
@@ -1,5 +1,0 @@
-package com.octagram.pollet.member.domain.model;
-
-public enum AuthProvider {
-	KAKAO, GOOGLE
-}
