--- src/main/java/com/octagram/pollet/member/domain/model/Rank.java
@@ -1,5 +1,0 @@
-package com.octagram.pollet.member.domain.model;
-
-public enum Rank {
-	BRONZE, SILVER, GOLD
-}
