+++ src/main/java/com/octagram/pollet/member/domain/type/AuthProvider.java
@@ -1,0 +1,5 @@
+package com.octagram.pollet.member.domain.type;
+
+public enum AuthProvider {
+	KAKAO, GOOGLE
+}
