+++ src/main/java/com/octagram/pollet/member/domain/type/MemberStatus.java
@@ -1,4 +1,4 @@
-package com.octagram.pollet.member.domain.model;
+package com.octagram.pollet.member.domain.type;
 
 public enum MemberStatus {
 	ACTIVE, SUSPENDED, BANNED
