+++ src/main/java/com/octagram/pollet/member/domain/type/Rank.java
@@ -1,0 +1,5 @@
+package com.octagram.pollet.member.domain.type;
+
+public enum Rank {
+	BRONZE, SILVER, GOLD
+}
