+++ src/main/java/com/octagram/pollet/member/domain/type/Role.java
@@ -1,0 +1,5 @@
+package com.octagram.pollet.member.domain.type;
+
+public enum Role {
+	GUEST, MEMBER, ADMIN
+}
